Remove left overs of initial experiments with liveview
