--- lib/coffee_web/components/ui/helpers/js_helpers.ex
@@ -1,34 +1,0 @@
-defmodule CoffeeWeb.UI.Helpers.JSHelpers do
-  alias Phoenix.LiveView.JS
-
-  @moduledoc """
-  Helper functions for JavaScript.
-  """
-  def show(js \\ %JS{}, selector) do
-    JS.show(js,
-      to: selector,
-      time: 300,
-      transition:
-        {"transition-all transform ease-out duration-300",
-         "opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95",
-         "opacity-100 translate-y-0 sm:scale-100"}
-    )
-  end
-
-  def hide(js \\ %JS{}, selector) do
-    JS.hide(js,
-      to: selector,
-      time: 200,
-      transition:
-        {"transition-all transform ease-in duration-200",
-         "opacity-100 translate-y-0 sm:scale-100",
-         "opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95"}
-    )
-  end
-
-  defmacro __using__(_) do
-    quote do
-      import CoffeeWeb.UI.Helpers.JSHelpers
-    end
-  end
-end
