--- lib/coffee_web/live/page_live.html.heex
@@ -1,31 +1,0 @@
-<%= @number %>
-<.button phx-click="add">Add</.button>
-<.simple_form for={@form} phx-submit="adding_more" class="mt-9">
-  <.input field={@form[:add_amount]} value={@form.params["add_amount"]} />
-  <.button class="w-full">Add More</.button>
-</.simple_form>
-<hr class="my-9" />
-
-<.button phx-click="confirm_modal">Confirm</.button>
-Open: <%= @modal_open %>
-
-<CoffeeWeb.UI.Modal.modal
-  :if={@modal_open}
-  show
-  id="confirm"
-  on_cancel={JS.push("close_modal")}
->
-  <h2>This is a modal.</h2>
-  <.button phx-click="close_modal">
-    Close Modal
-  </.button>
-</CoffeeWeb.UI.Modal.modal>
-
-<hr class="my-9" />
-
-<.alert variant="default" icon="hero-exclamation-circle">
-  <.alert_title>
-    Alert title
-  </.alert_title>
-  <.alert_description>Alert description</.alert_description>
-</.alert>
